+++ src/lib/tools/export.ts
@@ -3,10 +3,9 @@ import generateLinks from './links';
 import * as pure from './pure';
 import * as replace from './replace';
 import stats from './stats';
-import { timeNow, convertTime } from './time';
+import { timeNow, convertTime, uptime } from './time';
 import valueDiff from './valueDiff';
 import listItems from './summarizeItems';
 import summarize from './summarizeOffer';
-import uptime from './uptime';
 
 export { check, generateLinks, pure, replace, stats, timeNow, valueDiff, listItems, summarize, uptime, convertTime };
